Unified tabbed Settings screen (#112, #119)
Replace SettingsScreen + AssistantSettingsSheet with a single tabbed
settings screen (General, Instances, Agent, Tools). LLM provider config
moves from assistant bottom sheet to the Agent tab, MCP servers move to
the Tools tab. AssistantViewModel slimmed down — config methods moved to
SettingsViewModel, reactive sync via activeConfigFlow.
