+++ app/src/main/kotlin/io/github/leogallego/ansiblejane/assistant/data/IAssistantRepository.kt
@@ -1,6 +1,7 @@
 package io.github.leogallego.ansiblejane.assistant.data
 
 import io.github.leogallego.ansiblejane.assistant.engine.ChatMessage
+import kotlinx.coroutines.flow.Flow
 
 interface IAssistantRepository {
     fun addMessage(message: ChatMessage)
@@ -10,4 +11,5 @@ interface IAssistantRepository {
     suspend fun loadLlmConfig(): LlmProviderConfig?
     suspend fun saveAllLlmConfigs(configs: Map<String, LlmProviderConfig>)
     suspend fun loadAllLlmConfigs(): Map<String, LlmProviderConfig>
+    val activeConfigFlow: Flow<LlmProviderConfig?>
 }
