+++ app/src/main/kotlin/io/github/leogallego/ansiblejane/assistant/ui/AssistantScreen.kt
@@ -70,10 +70,10 @@ import org.koin.compose.viewmodel.koinViewModel
 
 @Composable
 fun AssistantScreen(
+    onNavigateToSettings: () -> Unit = {},
     viewModel: AssistantViewModel = koinViewModel()
 ) {
     val uiState by viewModel.uiState.collectAsStateWithLifecycle()
-    var showSettings by remember { mutableStateOf(false) }
 
     when (val state = uiState) {
         is AssistantUiState.Idle -> {
@@ -109,7 +109,7 @@ fun AssistantScreen(
             ActiveChatContent(
                 state = state,
                 onSendMessage = { viewModel.sendMessage(it) },
-                onOpenSettings = { showSettings = true },
+                onNavigateToSettings = onNavigateToSettings,
                 modifier = Modifier.fillMaxSize()
             )
         }
@@ -127,34 +127,6 @@ fun AssistantScreen(
             }
         }
     }
-
-    if (showSettings) {
-        val activeInstance = viewModel.activeInstance
-        val connections = (uiState as? AssistantUiState.Active)?.connections ?: emptyMap()
-        val currentLlmConfig by viewModel.llmConfig.collectAsStateWithLifecycle()
-        val allSavedConfigs by viewModel.savedConfigs.collectAsStateWithLifecycle()
-        val fetchedModels by viewModel.fetchedModels.collectAsStateWithLifecycle()
-        val modelFetchState by viewModel.modelFetchState.collectAsStateWithLifecycle()
-        AssistantSettingsSheet(
-            mcpEnabled = activeInstance?.mcpEnabled ?: false,
-            mcpServers = activeInstance?.mcpServerUrls ?: emptyList(),
-            connections = connections,
-            currentLlmConfig = currentLlmConfig,
-            savedConfigs = allSavedConfigs,
-            onToggleMcp = { viewModel.toggleMcpEnabled(it) },
-            onAddMcpServer = { url, label -> viewModel.addMcpServer(url, label) },
-            onRemoveMcpServer = { viewModel.removeMcpServer(it) },
-            onToggleReadOnly = { url, readOnly -> viewModel.toggleServerReadOnly(url, readOnly) },
-            fetchedModels = fetchedModels,
-            modelFetchState = modelFetchState,
-            onFetchModels = { url, key -> viewModel.fetchAvailableModels(url, key) },
-            onClearFetchedModels = { viewModel.clearFetchedModels() },
-            onSaveLlmConfig = { viewModel.updateLlmConfig(it) },
-            onSaveAllConfigs = { viewModel.updateAllLlmConfigs(it) },
-            onClearHistory = { viewModel.clearHistory() },
-            onDismiss = { showSettings = false }
-        )
-    }
 }
 
 @OptIn(ExperimentalLayoutApi::class)
@@ -162,7 +134,7 @@ fun AssistantScreen(
 private fun ActiveChatContent(
     state: AssistantUiState.Active,
     onSendMessage: (String) -> Unit,
-    onOpenSettings: () -> Unit,
+    onNavigateToSettings: () -> Unit,
     modifier: Modifier = Modifier
 ) {
     val listState = rememberLazyListState()
@@ -217,7 +189,7 @@ private fun ActiveChatContent(
                     style = MaterialTheme.typography.bodySmall,
                     color = MaterialTheme.colorScheme.onSurfaceVariant
                 )
-                TextButton(onClick = onOpenSettings, modifier = Modifier.testTag("button_configure")) {
+                TextButton(onClick = onNavigateToSettings, modifier = Modifier.testTag("button_configure")) {
                     Text("Configure", style = MaterialTheme.typography.bodySmall)
                 }
             }
@@ -228,7 +200,7 @@ private fun ActiveChatContent(
                     .padding(horizontal = 16.dp, vertical = 4.dp),
                 horizontalArrangement = Arrangement.End
             ) {
-                TextButton(onClick = onOpenSettings, modifier = Modifier.testTag("button_configure_llm")) {
+                TextButton(onClick = onNavigateToSettings, modifier = Modifier.testTag("button_configure_llm")) {
                     Text("Configure LLM", style = MaterialTheme.typography.bodySmall)
                 }
             }
