+++ app/src/main/kotlin/io/github/leogallego/ansiblejane/navigation/AppNavigation.kt
@@ -196,6 +196,9 @@ fun AppNavigation(
                     },
                     onNavigateToWorkflowJobStatus = { workflowJobId ->
                         navController.navigate(Routes.workflowJobStatus(workflowJobId))
+                    },
+                    onNavigateToSettings = {
+                        navController.navigate(Routes.SETTINGS)
                     }
                 )
             }
