+++ app/src/main/kotlin/io/github/leogallego/ansiblejane/presentation/settings/SettingsUiState.kt
@@ -1,17 +1,38 @@
 package io.github.leogallego.ansiblejane.presentation.settings
 
+import io.github.leogallego.ansiblejane.assistant.data.LlmProviderConfig
+import io.github.leogallego.ansiblejane.assistant.presentation.ModelFetchState
 import io.github.leogallego.ansiblejane.model.AapInstance
+import io.github.leogallego.ansiblejane.model.McpServerConfig
+import io.github.leogallego.ansiblejane.network.mcp.McpConnectionState
 import io.github.leogallego.ansiblejane.ui.components.TimeFormat
 
+enum class SettingsTab(val label: String) {
+    General("General"),
+    Instances("Instances"),
+    Agent("Agent"),
+    Tools("Tools")
+}
+
 sealed interface SettingsUiState {
-    data object Idle : SettingsUiState
     data object Loading : SettingsUiState
-    data class Success(
-        val instances: List<AapInstance>,
-        val selectedInstance: AapInstance?,
+    data class Ready(
+        val currentTab: SettingsTab = SettingsTab.General,
+        // Instances
+        val instances: List<AapInstance> = emptyList(),
+        val selectedInstance: AapInstance? = null,
         val selectedInstanceForDetails: AapInstance? = null,
+        // General
         val timezoneId: String? = null,
-        val timeFormat: TimeFormat = TimeFormat.SYSTEM
+        val timeFormat: TimeFormat = TimeFormat.SYSTEM,
+        // Agent (LLM config)
+        val savedConfigs: Map<String, LlmProviderConfig> = emptyMap(),
+        val activeConfig: LlmProviderConfig? = null,
+        val fetchedModels: List<String> = emptyList(),
+        val modelFetchState: ModelFetchState = ModelFetchState.Idle,
+        // Tools (MCP)
+        val mcpEnabled: Boolean = false,
+        val mcpServers: List<McpServerConfig> = emptyList(),
+        val connections: Map<String, McpConnectionState> = emptyMap()
     ) : SettingsUiState
-    data class Error(val message: String) : SettingsUiState
 }
